requirements.txt
scipy
numpy
Cython
scikit-image
opencv-python-headless
Pillow
matplotlib
imgaug>=0.4
albumentations>0.5
graphviz
tqdm
pyyaml
easydict
tensorboard
future
cffi
ninja
torch>=1.4.0
torchvision>=0.5.0
